backend/requirements.txt
# FastAPI Task Runner Requirements
# Core FastAPI and web framework dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database dependencies
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Redis and Celery for background tasks
redis>=5.0.1
celery>=5.3.4

# WebSocket support
websockets>=12.0

# HTTP client dependencies
httpx>=0.25.2
requests>=2.31.0

# OpenAI CUA Task Runner Requirements
# Core OpenAI and CUA dependencies
openai>=1.102.0
python-dotenv>=1.0.1

# Anthropic CUA Task Runner Requirements
# Core Anthropic dependencies
anthropic>=0.40.0

# Gemini CUA Task Runner Requirements
# Core Google Gemini dependencies
google-genai>=1.0.0

# Web automation and browser control
playwright>=1.55.0
browserbase>=1.2.0
scrapybara>=2.3.6
selenium>=4.15.0
webdriver-manager>=4.0.0

# HTTP and API handling
httpcore>=1.0.7

# Data processing and utilities
pandas>=1.3.0
pillow>=11.1.0
imagehash>=4.3.1
jsonpath-ng>=1.6.0

# Async and concurrency
anyio>=4.8.0
pyee>=12.1.1
greenlet>=3.1.1
nest-asyncio>=1.6.0

# Type hints and validation
annotated-types>=0.7.0
typing_extensions>=4.12.2
typing-inspection>=0.4.0

# HTTP utilities
h11>=0.14.0
idna>=3.10
charset-normalizer>=3.4.1
certifi>=2025.1.31
urllib3>=2.3.0
sniffio>=1.3.1
distro>=1.9.0
jiter>=0.8.2

# Additional utilities
pathlib2>=2.3.0
tqdm>=4.67.0
beautifulsoup4>=4.12.0
zipstream-ng>=1.1.1

# Docker container management
docker>=6.1.0
aiohttp>=3.9.0

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-httpx>=0.27.0
pytest-mock>=3.12.0
pytest-playwright>=0.5.0
freezegun>=1.4.0
faker>=24.0.0
fakeredis>=2.21.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Monitoring and logging
structlog>=23.2.0
rich>=13.7.0

# Authentication and OAuth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
email-validator>=2.0.0
flower>=2.0.0
openpyxl>=3.1.2
